Changes:

- allow the users to specify `IlcOptimizationPreference`, use `Speed` as default
- allow the users to specify `IlcInstructionSet`, use host process settings as default

Sample benchmark:

```cs
private string a = new string('a', 1000) + "a";
private string b = new string('a', 1000) + "b";

[Benchmark]
public bool Compare() => a.Equals(b, System.StringComparison.Ordinal);
```

```cmd
dotnet run -c Release -f net6.0 --filter *Compare --job short --keepFiles true --runtimes nativeaot7.0
```

## Before

|  Method |       Runtime |          Toolchain |     Mean |    Error |   StdDev |
|-------- |-------------- |------------------- |---------:|---------:|---------:|
| Compare | NativeAOT 7.0 | ILCompiler 7.0.0-* | 66.88 ns | 2.263 ns | 0.124 ns |

## After

|  Method |       Runtime |          Toolchain |     Mean |    Error |   StdDev |
|-------- |-------------- |------------------- |---------:|---------:|---------:|
| Compare | NativeAOT 7.0 | ILCompiler 7.0.0-* | 35.42 ns | 0.539 ns | 0.030 ns |

Sample settings on my x64 AMD machine:

```xml
<ItemGroup>
    <IlcArg Include="--instructionset:base,sse,sse2,sse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt" />
</ItemGroup>
```




cc @jkotas @MichalStrehovsky